more refactoring work
