+++ quickannotator/api/v1/annotation/__init__.py
@@ -16,7 +16,7 @@ from .helper import (
 )
 from datetime import datetime
 from quickannotator.api.v1.tile.helper import upsert_tile, point_to_tileid, tile_intersects_mask
-from quickannotator.api.v1.image.helper import get_image_by_id
+from quickannotator.api.v1.image.utils import get_image_by_id
 from quickannotator.api.v1.annotation_class.helper import get_annotation_class_by_id
 from quickannotator.api.v1.utils.shared_crud import insert_new_annotation, get_annotation_query
